Move SmileyView from App module to SmileRating module   I want SmileyView because my project showed SmileyView   I think this is not just my project.
