+++ app/src/main/res/layout/activity_test.xml
@@ -3,7 +3,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.example.smilyrating.SmileyView
+    <com.hsalf.smileyrating.SmileyView
+        android:id="@+id/smiley_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_centerInParent="true" />
